create Components for StyledComponents of PrewiewCardList
