+++ components/StyledLink/index.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+import Link from "next/link";
+
+export const StyledLink = styled(Link)`
+  text-decoration: none;
+  padding: 0.5rem 1rem;
+  color: black;
+  background-color: #f2d5a3;
+  border-radius: 4px;
+  cursor: pointer;
+`;
