+++ components/StyledList/index.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const StyledList = styled.ul`
+  list-style-type: none;
+  padding: 0;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1rem;
+`;
