+++ components/StyledListItem/index.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const StyledListItem = styled.li`
+  border: 2px solid black;
+  border-radius: 1rem;
+  width: 90%;
+  position: relative;
+`;
